Always snap line crosshair to nearest point
Previously, the line crosshair would only snap to the nearest point if
it was within 20 pixels.  This commit changes that behavior, so that the
line crosshair will snap to the nearest point, no matter how near it is.

This provides a nicer user experience. It also lets us fine tune the
spatial index used to look up hover points. In particular, since the
`INDEX_BIN_WIDTH` constant is no longer coupled to the user experience
of how much the crosshair will snap, we can change it at will. This
commit shrinks that constant, which should speed up reads from the
index.