Fix tooltip flickering
Fixes an issue where moving the mouse rapidly towards the tooltip would
cause the tooltip to disappear momentarily and then reappear.

The fix involved using the `mouseenter` / `mouseover` / `mouseleave` set
of events to read the approriate mouse hover position, even if the mouse
was over the tooltip rather than the plot interaction region. This in
turn required:

1. Making the tooltip a descendent of the plot interaction region in the
   React render tree
2. Using React's `onMouseEnter`, `onMouseOver`, and `onMouseLeave`
   events rather than adding event listeners to the plot interaction
   region directly. This is because the tooltip is rendered using a
   React portal, which has [special handling for event bubbling][0].

[0]: https://reactjs.org/docs/portals.html#event-bubbling-through-portals