Alter value formatting public API
Previously a consumer of the vis library could supply optional
`xTickFormatter` and `yTickFormatter` fields in their config:

    {
       ...someConfigOptions,
       xTickFormatter: t => new Date(t).getFullYear(),
       yTickFormatter: t => `${Math.round(t)}%`
    }

This commit replaces these options with a `valueFormatters` option,
which is an object. Each key in the object is a column key, and each
value in the object is a function that specifies how to format values
within that column.

Closes #37
Closes #40