Update README.md (#292)
We create a table with 5 rows, but only give it 3 sets of values. Either we need to add two more sets of values (as is done here: https://github.com/influxdata/giraffe/blob/master/giraffe/README.md#config) or reduce the number of rows to 3. I took the easier approach (reducing to 3).